chore: uplot file renames and reference updates (#4004)
