+++ frontend/src/lib/uPlotLib/getUplotChartOptions.ts
