+++ frontend/src/lib/uPlotLib/utils/getUplotChartData.ts
